authentication almost complete

+++ src/api/services/db.js
@@ -18,9 +18,25 @@ const createUser = async (displayName, email, password) => {
             });
         }
         const hashedPassword = await encryptPassword(password);
-        const result = await usersCollection.insertOne({ display_name: displayName, email, password: hashedPassword, verified: false });
+        const result = await usersCollection.insertOne({ display_name: displayName, email, password: hashedPassword, tokens: 3, verified: false });
         await client.close();
         return new User(result.insertedId, displayName, email);
+    }
+    catch (err) {
+        throw new ErrorObject({
+            message: err.message,
+            statusCode: err.statusCode
+        });
+    }
+}
+
+const updatePassword = async (user_id, password) => {
+    try {
+        const client = await MongoClient.connect(db.url);
+        const usersCollection = client.db(db.name).collection('users');
+        const hashedPassword = await encryptPassword(password);
+        await usersCollection.updateOne({ _id: ObjectId(user_id) }, { $set: { password: hashedPassword } });
+        await client.close();
     }
     catch (err) {
         throw new ErrorObject({
@@ -76,11 +92,11 @@ const getUserById = async (user_id) => {
     }
 }
 
-const setRefreshToken = async (user_id, refreshToken) => {
+const createRefreshToken = async (user_id, refreshToken) => {
     try {
         const client = await MongoClient.connect(db.url);
         const refreshTokensCollection = client.db(db.name).collection('refresh_tokens');
-        await refreshTokensCollection.updateOne({ user_id }, { $set: { refresh_token: refreshToken } }, { upsert: true });
+        await refreshTokensCollection.insertOne({ user_id, refresh_token: refreshToken });
         await client.close();
     }
     catch (err) {
@@ -91,11 +107,26 @@ const setRefreshToken = async (user_id, refreshToken) => {
     }
 }
 
-const getRefreshToken = async (user_id) => {
+const setRefreshToken = async (user_id, refreshToken) => {
     try {
         const client = await MongoClient.connect(db.url);
         const refreshTokensCollection = client.db(db.name).collection('refresh_tokens');
-        const tokenObject = await refreshTokensCollection.findOne({ user_id: ObjectId(user_id) });
+        await refreshTokensCollection.updateOne({ user_id, refresh_token: refreshToken }, { $set: { refresh_token: refreshToken } }, { upsert: true });
+        await client.close();
+    }
+    catch (err) {
+        throw new ErrorObject({
+            message: err.message,
+            statusCode: err.statusCode
+        });
+    }
+}
+
+const getRefreshToken = async (refreshToken) => {
+    try {
+        const client = await MongoClient.connect(db.url);
+        const refreshTokensCollection = client.db(db.name).collection('refresh_tokens');
+        const tokenObject = await refreshTokensCollection.findOne({ refresh_token: refreshToken });
         if (!tokenObject) {
             await client.close();
             throw new ErrorObject({
@@ -114,11 +145,26 @@ const getRefreshToken = async (user_id) => {
     }
 }
 
-const deleteRefreshToken = async (refreshToken) => {
+const deleteRefreshToken = async (user_id, refreshToken) => {
     try {
         const client = await MongoClient.connect(db.url);
         const refreshTokensCollection = client.db(db.name).collection('refresh_tokens');
-        await refreshTokensCollection.deleteOne({ refresh_token: refreshToken });
+        await refreshTokensCollection.deleteOne({ user_id, refresh_token: refreshToken });
+        await client.close();
+    }
+    catch (err) {
+        throw new ErrorObject({
+            message: err.message,
+            statusCode: err.statusCode
+        });
+    }
+}
+
+const deleteRefreshTokens = async (user_id) => {
+    try {
+        const client = await MongoClient.connect(db.url);
+        const refreshTokensCollection = client.db(db.name).collection('refresh_tokens');
+        await refreshTokensCollection.deleteMany({ user_id });
         await client.close();
     }
     catch (err) {
@@ -133,7 +179,14 @@ const verifyUser = async (user_id) => {
     try {
         const client = await MongoClient.connect(db.url);
         const usersCollection = client.db(db.name).collection('users');
-        await usersCollection.updateOne({ user_id }, { $set: { verified: true } });
+        const user = await usersCollection.findOne({ _id: ObjectId(user_id) });
+        if (user.verified) {
+            throw new ErrorObject({
+                message: 'Email already verified',
+                statusCode: 400
+            });
+        }
+        await usersCollection.updateOne({ _id: ObjectId(user_id) }, { $set: { verified: true } });
         await client.close();
     }
     catch (err) {
@@ -146,10 +199,13 @@ const verifyUser = async (user_id) => {
 
 module.exports = {
     createUser,
+    updatePassword,
     getUserByEmail,
     getUserById,
+    createRefreshToken,
     setRefreshToken,
     getRefreshToken,
     deleteRefreshToken,
+    deleteRefreshTokens,
     verifyUser
 }
